Editor: Move rope_patch.py to utils/introspection to match jedi_patch.py
