+++ spyder/utils/introspection/rope_patch.py
